+++ packages/scss-config-webpack-plugin/src/config/production.config.ts
@@ -2,12 +2,11 @@ import { FrontlineScssWebpackPluginOptions } from "./FrontlineScssConfigWebpackP
 
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const OptimizeCSSAssetsPlugin = require("optimize-css-assets-webpack-plugin");
-const autoprefixer = require("autoprefixer");
-const jsonImporter = require("node-sass-json-importer");
-const globImporter = require("node-sass-glob-importer");
 const safeParser = require("postcss-safe-parser");
 
 import { FILE_MODULE_REGEX, FILE_REGEX } from "../utils/file-extension";
+import { FrontlinePostcssConfig } from "./FrontlinePostcssConfig";
+import { FrontlineNodeSassConfig } from "./FrontlineNodeSassConfig";
 
 export = (options: FrontlineScssWebpackPluginOptions) => {
     return {
@@ -18,7 +17,9 @@ export = (options: FrontlineScssWebpackPluginOptions) => {
                     exclude: FILE_MODULE_REGEX,
                     use: [
                         {
-                            loader: MiniCssExtractPlugin.loader
+                            loader: MiniCssExtractPlugin.loader,
+                            // css is located in `static/css`, use '../../' to locate index.html folder
+                            options: { publicPath: options.publicPath }
                         },
                         {
                             loader: require.resolve("css-loader"),
@@ -28,21 +29,18 @@ export = (options: FrontlineScssWebpackPluginOptions) => {
                         },
                         {
                             loader: require.resolve("postcss-loader"),
-                            options: {
-                                plugins: (loader: any) => [
-                                    require("postcss-flexbugs-fixes"),
-                                    autoprefixer({
-                                        // flexbox: "no-2009" will add prefixes only for final and IE versions of specification.
-                                        // @see https://github.com/postcss/autoprefixer#disabling
-                                        flexbox: "no-2009",
-                                        env: options.browserslistEnv
-                                    }),
-                                    require("iconfont-webpack-plugin")({
-                                        resolve: loader.resolve
-                                    })
-                                ],
-                                sourceMap: false
-                            }
+                            options:
+                                options.postCssConfigFile !== null
+                                    ? {
+                                          postcssOptions: {
+                                              path: options.postCssConfigFile
+                                          }
+                                      }
+                                    : {
+                                          postcssOptions: FrontlinePostcssConfig(
+                                              options.browserslistEnv
+                                          )
+                                      }
                         },
                         {
                             loader: require.resolve("resolve-url-loader"),
@@ -52,12 +50,9 @@ export = (options: FrontlineScssWebpackPluginOptions) => {
                         },
                         {
                             loader: require.resolve("sass-loader"),
-                            options: {
-                                sourceMap: true,
-                                sassOptions: {
-                                    importer: [jsonImporter(), globImporter()]
-                                }
-                            }
+                            options: FrontlineNodeSassConfig(
+                                options.sassOptions
+                            )
                         }
                     ]
                 },
@@ -65,7 +60,9 @@ export = (options: FrontlineScssWebpackPluginOptions) => {
                     test: FILE_MODULE_REGEX,
                     use: [
                         {
-                            loader: MiniCssExtractPlugin.loader
+                            loader: MiniCssExtractPlugin.loader,
+                            // css is located in `static/css`, use '../../' to locate index.html folder
+                            options: { publicPath: options.publicPath }
                         },
                         {
                             loader: require.resolve("css-loader"),
@@ -76,21 +73,18 @@ export = (options: FrontlineScssWebpackPluginOptions) => {
                         },
                         {
                             loader: require.resolve("postcss-loader"),
-                            options: {
-                                plugins: (loader: any) => [
-                                    require("postcss-flexbugs-fixes"),
-                                    autoprefixer({
-                                        // flexbox: "no-2009" will add prefixes only for final and IE versions of specification.
-                                        // @see https://github.com/postcss/autoprefixer#disabling
-                                        flexbox: "no-2009",
-                                        env: options.browserslistEnv
-                                    }),
-                                    require("iconfont-webpack-plugin")({
-                                        resolve: loader.resolve
-                                    })
-                                ],
-                                sourceMap: false
-                            }
+                            options:
+                                options.postCssConfigFile !== null
+                                    ? {
+                                          postcssOptions: {
+                                              path: options.postCssConfigFile
+                                          }
+                                      }
+                                    : {
+                                          postcssOptions: FrontlinePostcssConfig(
+                                              options.browserslistEnv
+                                          )
+                                      }
                         },
                         {
                             loader: require.resolve("resolve-url-loader"),
@@ -100,12 +94,9 @@ export = (options: FrontlineScssWebpackPluginOptions) => {
                         },
                         {
                             loader: require.resolve("sass-loader"),
-                            options: {
-                                sourceMap: true,
-                                sassOptions: {
-                                    importer: [jsonImporter(), globImporter()]
-                                }
-                            }
+                            options: FrontlineNodeSassConfig(
+                                options.sassOptions
+                            )
                         }
                     ]
                 }