+++ packages/font-config-webpack-plugin/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.9.0-rc.0](https://github.com/akqa-frontline/frontline/compare/v0.8.1...v0.9.0-rc.0) (2020-06-11)
+
+
+### Features
+
+* **generate-injection-html-webpack-plugin:** generate "partial" html files for injection and loadin ([b7ab71b](https://github.com/akqa-frontline/frontline/commit/b7ab71b6386bb103570b0dddf91a6aef46f19e87))
+
+
+
+
+
+
 ## [0.7.2-alpha.0](https://github.com/akqa-frontline/frontline/compare/v0.7.1...v0.7.2-alpha.0) (2020-05-22)
 
 **Note:** Version bump only for package @akqa-frontline/font-config-webpack-plugin